Fixed LevelSelection
Fixed level selection so that unchecking the booleans will disable the
locked images,

NOTE: You will have to refresh by disabling and enabling the level
selectpanel. It does not refresh on Update(). In otherwords, uncheck the
booleans/levels that you so wish, and then disable, and reenable
LevelSelect panel. The images will then udate.