created levels 1 through 5
I made the first 5 levels. the first level contains no blocks so its a
straight shot to the goal. The second level contains a few blocks but it
is able to be completed with a positive ion. The third level is a
tougher version of level two. The fourth level has a wall of blocks that
the player shoots between and uses an negatiive ion to direct the
electorn to the goal. The fith level tests the player's skill to beat
the fith level.